move needle react to react generator
