+++ generators/base/generator-base.mjs
@@ -596,30 +596,6 @@ export default class JHipsterBaseGenerator extends PrivateBase {
 
   /**
    * @private
-   * Add new scss style to the react application in "app.scss".
-   *
-   * @param {string} style - css to add in the file
-   * @param {string} comment - comment to add before css code
-   *
-   * example:
-   *
-   * style = '.jhipster {\n     color: #baa186;\n}'
-   * comment = 'New JHipster color'
-   *
-   * * ==========================================================================
-   * New JHipster color
-   * ========================================================================== *
-   * .jhipster {
-   *     color: #baa186;
-   * }
-   *
-   */
-  addAppSCSSStyle(style, comment) {
-    this.needleApi.clientReact.addAppSCSSStyle(style, comment);
-  }
-
-  /**
-   * @private
    * Copy third-party library resources path.
    *
    * @param {string} sourceFolder - third-party library resources source path
