+++ generators/needle-api.mjs
@@ -19,7 +19,7 @@
 import Base from './needle-base.mjs';
 import Client from './client/needle-api/needle-client.mjs';
 import ClientAngular from './angular/needle-api/needle-client-angular.mjs';
-import ClientReact from './client/needle-api/needle-client-react.mjs';
+import ClientReact from './react/needle-api/needle-client-react.mjs';
 import ClientVue from './client/needle-api/needle-client-vue.mjs';
 import ClientWebpack from './client/needle-api/needle-client-webpack.mjs';
 import ClientI18n from './client/needle-api/needle-client-i18n.mjs';
