+++ generators/react/needle-api/needle-client-react.mjs
@@ -19,7 +19,7 @@
 import chalk from 'chalk';
 import _ from 'lodash';
 
-import needleClientBase from './needle-client.mjs';
+import needleClientBase from '../../client/needle-api/needle-client.mjs';
 import { stripMargin } from '../../base/support/index.mjs';
 
 export default class extends needleClientBase {
